Improve datastore later implementation.
Our previous implementation forced us to stick with DTab files (which is
techincally ideal, especially in a production environment), however, we
run into a setup overhead for a new/curious developer. By leaving the
data-layer sticking to c.t.app.App flags, we can heavily reduce the
amount of setup ovehead associated.

This includes ease of use changes as well.